refactor: update book api
