+++ src/redux/services/books.ts
@@ -1,41 +1,44 @@
-import { baseApi } from './baseApi';
+import { baseNewApi } from './baseNewApi';
 import {
-  BookById,
-  BookByIdResponse,
-  BooksArgs,
-  BooksListResponse,
+  BookByIdNew,
+  BooksArgsNew,
+  BooksMainPage,
+  BooksResponse,
 } from './services.types';
-
-export const booksApi = baseApi.injectEndpoints({
+export const booksApi = baseNewApi.injectEndpoints({
   endpoints: (builder) => ({
-    getBooks: builder.query<BooksListResponse, BooksArgs>({
+    getBooks: builder.query<BooksResponse, BooksArgsNew>({
       query: ({
-        page = 1,
-        size = 9,
-        sort = [''],
-        search = '',
-        price = [],
-        language = [],
-        categories = [],
+        lang,
         years = [],
-        authors = [],
+        languages = [],
+        minPrice = '',
+        maxPrice = '',
+        sortPrice = 'ASC',
+        isNew = false,
+        alphabetical = false,
+        limit = 16,
+        page = 1,
       }) => ({
-        url: `/open/book/list?page=${page}&size=${size}&sort=${sort}&title=${search}&price=${price.join(',')}&language=${language.join(',')}&year=${years.join(',')}&category=${categories.join(',')}&author=${authors.join(',')}`,
+        url: `book?lang=${lang}&page=${page}&limit=${limit}&years=${years.join(',')}&languages=${languages.join(',')}&minPrice=${minPrice}&maxPrice=${maxPrice}&sortPrice=${sortPrice}&new=${isNew}&alphabetical=${alphabetical}`,
       }),
-      providesTags: ['Books'],
     }),
-    getBookById: builder.query<BookById, string>({
-      query: (id) => ({
-        url: `/open/book/${id}`,
+    getBooksForMain: builder.query<BooksMainPage, { lang: string }>({
+      query: ({ lang }) => ({
+        url: `/book/main-page?lang=${lang}`,
       }),
-      providesTags: ['Books'],
-      transformResponse: (response: BookByIdResponse) => ({
-        ...response,
-        year: [{ id: response.year, name: String(response.year) }],
+    }),
+    getBookById: builder.query<BookByIdNew, { id: string; lang: string }>({
+      query: ({ id, lang }) => ({
+        url: `/book/${id}?lang=${lang}`,
       }),
     }),
   }),
   overrideExisting: false,
 });
 
-export const { useGetBooksQuery, useGetBookByIdQuery } = booksApi;
+export const {
+  useGetBooksQuery,
+  useGetBookByIdQuery,
+  useGetBooksForMainQuery,
+} = booksApi;
